+++ src/application-insights/azext_applicationinsights/_client_factory.py
@@ -36,7 +36,7 @@ def cf_events(cli_ctx, _):
 
 
 def cf_components(cli_ctx, _):
-    return applicationinsights_mgmt_plane_client(cli_ctx).components
+    return applicationinsights_mgmt_plane_client(cli_ctx, api_version='2018-05-01-preview').components
 
 
 def cf_component_billing(cli_ctx, _):